validator: ensure user doesn't try to mount /sys without userns
Inside a user namespace, mount permissions become more nuanced. sysfs
requires you to have either CAP_SYS_ADMIN in the root user namespace
(for us this means that we're not using a user namespace) or that you
have CAP_SYS_ADMIN in the user namespace the network namespace was
created in. This means that having just a user namespace and no private
network namespace will result in errors when mounting sysfs[1]. Warn the
user about this in the validator.

[1]: http://lists.linuxfoundation.org/pipermail/containers/2013-August/033388.html

Signed-off-by: Aleksa Sarai <asarai@suse.de>