Use SELinux mount label if SELinux is enabled on host
We need to differentiate between SELinux disabled for containers
from SELinux disabled on the host. If SELinux is disabled inside
of containers we still want to use a default mount label. If it
is disabled on the host we don't want to use any labels.

SELinux policy is going to require all files to be labeled, in order
for docker and default containers to work correctly going forward.

These changes will set the default label to svirt_sandbox_file_t:s0.C1023
Which no container will be allowed to read/write. This label can be used
for spc/unconfined containers.

Signed-off-by: Dan Walsh dwalsh@redhat.com

Signed-off-by: Dan Walsh <dwalsh@redhat.com>