Update state JSON encoding fields to match spec
In #1048 the `cState` struct was consolidated into the `containerState`
struct which inadvertently changed the `bundlePath` and `rootfsPath`
fields in the JSON to `bundle` and `rootfs` respectively.

https://github.com/opencontainers/runtime-spec/blob/master/runtime.md#state
requires that the bundle path field be named `bundlePath` so reinstate
that name.

The specification does not cover the rootfs field name, but revert it
too for consistency.

This also changes the output of `runc list --format json` so adjust the
test case accordingly.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>