checkpoint: add pre-dump support
This is based on existing pull requests to implement pre-dump support in
runc. In addition to rebasing this tries to address the open questions
in the previous pull request.

One questions was why the code was no longer waiting (cmd.Process.Wait())
for CRIU in the pre-dump case. This is related to how CRIU expects a
pre-dump to work. Once a pre-dump is started the CRIU RPC process goes
into a loop where it either expects further pre-dumps or the final dump.
From CRIU's point of view this makes sense as it expects the user (runc)
to make as many pre-dumps as necessary and then the final dump.
In its final implementation that is probably also what runc should do.
In this first approach the incremental checkpoints are created manually
by runc as well as the final checkpoint. Therefore the code is still not
waiting for CRIU to end, because it never will. But if we get the RPC
message that the pre-dump was successful CRIU is just terminated.

Instead of passing the pre-dump as additional variable to Checkpoint()
the information about doing a pre-dump is now part of criuOpts.

If referencing a parent image for pre-dump mode the option TrackMem is
now passed to CRIU as it is required.

Examples:

 1. Dump:
  runc checkpoint <container-id> --pre-dump --image-path parent
 2. Final Dump:
  runc checkpoint <container-id> --image-path image --parent-path ../parent
 3: Restore:
  runc restore <container-id> --image-path image

In one test the initial dump was 182M.  The final dump only 6.6M which
decreased the downtime to the time required to transfer 6.6M instead of
182M.

Signed-off-by: Adrian Reber <areber@redhat.com>